+++ aws-sdk-core/lib/aws-sdk-core/xml/builder.rb
@@ -66,7 +66,7 @@ module Aws
         when 'structure' then structure(name, shape, value)
         when 'list'      then list(name, shape, value)
         when 'timestamp' then node(name, shape, shape.format_time(value, 'iso8601'))
-        when 'blob'      then node(name, shape, Base64.strict_encode64(value))
+        when 'blob'      then node(name, shape, Base64.strict_encode64(value.read))
         when 'map'       then raise NotImplementedError, 'not supported'
         else
           node(name, shape, value.to_s)